openai-compat fix
