+++ packages/openai-compatible/src/chat/openai-compatible-chat-language-model.ts
@@ -58,6 +58,13 @@ export type OpenAICompatibleChatConfig = {
    * The supported URLs for the model.
    */
   supportedUrls?: () => LanguageModelV3['supportedUrls'];
+
+  /**
+   * Optional function to transform the request body before sending it to the API.
+   * This is useful for proxy providers that may require a different request format
+   * than the official OpenAI API.
+   */
+  transformRequestBody?: (args: Record<string, any>) => Record<string, any>;
 };
 
 export class OpenAICompatibleChatLanguageModel implements LanguageModelV3 {
@@ -98,6 +105,10 @@ export class OpenAICompatibleChatLanguageModel implements LanguageModelV3 {
 
   get supportedUrls() {
     return this.config.supportedUrls?.() ?? {};
+  }
+
+  private transformRequestBody(args: Record<string, any>): Record<string, any> {
+    return this.config.transformRequestBody?.(args) ?? args;
   }
 
   private async getArgs({
@@ -219,7 +230,8 @@ export class OpenAICompatibleChatLanguageModel implements LanguageModelV3 {
   ): Promise<LanguageModelV3GenerateResult> {
     const { args, warnings } = await this.getArgs({ ...options });
 
-    const body = JSON.stringify(args);
+    const transformedBody = this.transformRequestBody(args);
+    const body = JSON.stringify(transformedBody);
 
     const {
       responseHeaders,
@@ -231,7 +243,7 @@ export class OpenAICompatibleChatLanguageModel implements LanguageModelV3 {
         modelId: this.modelId,
       }),
       headers: combineHeaders(this.config.headers(), options.headers),
-      body: args,
+      body: transformedBody,
       failedResponseHandler: this.failedResponseHandler,
       successfulResponseHandler: createJsonResponseHandler(
         OpenAICompatibleChatResponseSchema,
@@ -321,7 +333,7 @@ export class OpenAICompatibleChatLanguageModel implements LanguageModelV3 {
   ): Promise<LanguageModelV3StreamResult> {
     const { args, warnings } = await this.getArgs({ ...options });
 
-    const body = {
+    const body = this.transformRequestBody({
       ...args,
       stream: true,
 
@@ -329,7 +341,7 @@ export class OpenAICompatibleChatLanguageModel implements LanguageModelV3 {
       stream_options: this.config.includeUsage
         ? { include_usage: true }
         : undefined,
-    };
+    });
 
     const metadataExtractor =
       this.config.metadataExtractor?.createStreamExtractor();
