+++ packages/openai-compatible/src/openai-compatible-provider.ts
@@ -89,6 +89,13 @@ Include usage information in streaming responses.
    * Whether the provider supports structured outputs in chat models.
    */
   supportsStructuredOutputs?: boolean;
+
+  /**
+   * Optional function to transform the request body before sending it to the API.
+   * This is useful for proxy providers that may require a different request format
+   * than the official OpenAI API.
+   */
+  transformRequestBody?: (args: Record<string, any>) => Record<string, any>;
 }
 
 /**
@@ -146,6 +153,7 @@ export function createOpenAICompatible<
       ...getCommonModelConfig('chat'),
       includeUsage: options.includeUsage,
       supportsStructuredOutputs: options.supportsStructuredOutputs,
+      transformRequestBody: options.transformRequestBody,
     });
 
   const createCompletionModel = (modelId: COMPLETION_MODEL_IDS) =>
