+++ packages/openai/src/chat/openai-chat-language-model.test.ts
@@ -3493,145 +3493,4 @@ describe('doStream', () => {
       expect(chunks.filter(chunk => chunk.type === 'raw')).toHaveLength(0);
     });
   });
-
-  describe('transformRequestBody', () => {
-    function prepareTransformJsonResponse() {
-      server.urls['https://api.openai.com/v1/chat/completions'].response = {
-        type: 'json-value',
-        body: {
-          id: 'chatcmpl-test',
-          object: 'chat.completion',
-          created: 1711115037,
-          model: 'gpt-4',
-          choices: [
-            {
-              index: 0,
-              message: {
-                role: 'assistant',
-                content: 'Hello!',
-              },
-              finish_reason: 'stop',
-            },
-          ],
-          usage: {
-            prompt_tokens: 4,
-            total_tokens: 34,
-            completion_tokens: 30,
-          },
-        },
-      };
-    }
-
-    function prepareTransformStreamResponse() {
-      server.urls['https://api.openai.com/v1/chat/completions'].response = {
-        type: 'stream-chunks',
-        chunks: [
-          `data: {"id":"chatcmpl-test","object":"chat.completion.chunk","created":1711115037,"model":"gpt-4","choices":[{"index":0,"delta":{"role":"assistant","content":"Hello"},"finish_reason":null}]}\n\n`,
-          `data: {"id":"chatcmpl-test","object":"chat.completion.chunk","created":1711115037,"model":"gpt-4","choices":[{"index":0,"delta":{"content":"!"},"finish_reason":"stop"}],"usage":{"prompt_tokens":4,"completion_tokens":2,"total_tokens":6}}\n\n`,
-          'data: [DONE]\n\n',
-        ],
-      };
-    }
-
-    it('should transform request body in doGenerate when transformRequestBody is provided', async () => {
-      const transformFn = vi.fn((body: Record<string, any>) => ({
-        ...body,
-        custom_field: 'added-by-transform',
-      }));
-
-      prepareTransformJsonResponse();
-
-      const { OpenAIChatLanguageModel } = await import(
-        './openai-chat-language-model'
-      );
-      const model = new OpenAIChatLanguageModel('gpt-4', {
-        provider: 'test-provider',
-        url: ({ path }) => `https://api.openai.com/v1${path}`,
-        headers: () => ({}),
-        transformRequestBody: transformFn,
-      });
-
-      await model.doGenerate({
-        prompt: TEST_PROMPT,
-      });
-
-      // Verify transform was called
-      expect(transformFn).toHaveBeenCalledOnce();
-      expect(transformFn).toHaveBeenCalledWith(
-        expect.objectContaining({
-          model: 'gpt-4',
-          messages: [{ role: 'user', content: 'Hello' }],
-        }),
-      );
-
-      // Verify transformed body was sent
-      expect(await server.calls[0].requestBodyJson).toMatchObject({
-        custom_field: 'added-by-transform',
-      });
-    });
-
-    it('should transform request body in doStream when transformRequestBody is provided', async () => {
-      const transformFn = vi.fn((body: Record<string, any>) => ({
-        ...body,
-        custom_field: 'added-by-transform',
-      }));
-
-      prepareTransformStreamResponse();
-
-      const { OpenAIChatLanguageModel } = await import(
-        './openai-chat-language-model'
-      );
-      const model = new OpenAIChatLanguageModel('gpt-4', {
-        provider: 'test-provider',
-        url: ({ path }) => `https://api.openai.com/v1${path}`,
-        headers: () => ({}),
-        transformRequestBody: transformFn,
-      });
-
-      const { stream } = await model.doStream({
-        prompt: TEST_PROMPT,
-      });
-
-      // Consume the stream
-      await convertReadableStreamToArray(stream);
-
-      // Verify transform was called
-      expect(transformFn).toHaveBeenCalledOnce();
-      expect(transformFn).toHaveBeenCalledWith(
-        expect.objectContaining({
-          model: 'gpt-4',
-          messages: [{ role: 'user', content: 'Hello' }],
-          stream: true,
-        }),
-      );
-
-      // Verify transformed body was sent
-      expect(await server.calls[0].requestBodyJson).toMatchObject({
-        custom_field: 'added-by-transform',
-      });
-    });
-
-    it('should work without transformRequestBody', async () => {
-      prepareTransformJsonResponse();
-
-      const { OpenAIChatLanguageModel } = await import(
-        './openai-chat-language-model'
-      );
-      const model = new OpenAIChatLanguageModel('gpt-4', {
-        provider: 'test-provider',
-        url: ({ path }) => `https://api.openai.com/v1${path}`,
-        headers: () => ({}),
-      });
-
-      await model.doGenerate({
-        prompt: TEST_PROMPT,
-      });
-
-      const requestBody = await server.calls[0].requestBodyJson;
-      expect(requestBody).toMatchObject({
-        model: 'gpt-4',
-      });
-      expect(requestBody).not.toHaveProperty('custom_field');
-    });
-  });
 });
