+++ packages/openai/src/responses/openai-responses-language-model.test.ts
@@ -11,7 +11,7 @@ import {
 } from '@ai-sdk/provider-utils/test';
 import { createTestServer } from '@ai-sdk/test-server/with-vitest';
 import fs from 'node:fs';
-import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { beforeEach, describe, expect, it } from 'vitest';
 import { OpenAIResponsesLanguageModel } from './openai-responses-language-model';
 import {
   openaiResponsesModelIds,
@@ -6922,159 +6922,6 @@ describe('OpenAIResponsesLanguageModel', () => {
           },
         ]
       `);
-    });
-  });
-
-  describe('transformRequestBody', () => {
-    function prepareTransformJsonResponse() {
-      server.urls['https://api.openai.com/v1/responses'].response = {
-        type: 'json-value',
-        body: {
-          id: 'resp_test',
-          object: 'response',
-          created_at: 1711115037,
-          model: 'gpt-4o',
-          output: [
-            {
-              type: 'message',
-              id: 'msg_test',
-              role: 'assistant',
-              content: [
-                {
-                  type: 'output_text',
-                  text: 'Hello!',
-                  annotations: [],
-                },
-              ],
-            },
-          ],
-          usage: {
-            input_tokens: 4,
-            output_tokens: 30,
-          },
-        },
-      };
-    }
-
-    function prepareTransformStreamResponse() {
-      server.urls['https://api.openai.com/v1/responses'].response = {
-        type: 'stream-chunks',
-        chunks: [
-          `data: {"type":"response.created","response":{"id":"resp_test","object":"response","created_at":1711115037,"model":"gpt-4o","output":[],"usage":null}}\n\n`,
-          `data: {"type":"response.output_item.added","output_index":0,"item":{"type":"message","id":"msg_test","role":"assistant","content":[]}}\n\n`,
-          `data: {"type":"response.output_text.delta","output_index":0,"content_index":0,"item_id":"msg_test","delta":"Hello!"}\n\n`,
-          `data: {"type":"response.output_item.done","output_index":0,"item":{"type":"message","id":"msg_test","role":"assistant","content":[{"type":"output_text","text":"Hello!","annotations":[]}]}}\n\n`,
-          `data: {"type":"response.completed","response":{"id":"resp_test","object":"response","created_at":1711115037,"model":"gpt-4o","output":[{"type":"message","id":"msg_test","role":"assistant","content":[{"type":"output_text","text":"Hello!","annotations":[]}]}],"usage":{"input_tokens":4,"output_tokens":2}}}\n\n`,
-          'data: [DONE]\n\n',
-        ],
-      };
-    }
-
-    it('should transform request body in doGenerate when transformRequestBody is provided', async () => {
-      const transformFn = vi.fn((body: Record<string, any>) => ({
-        ...body,
-        custom_field: 'added-by-transform',
-      }));
-
-      prepareTransformJsonResponse();
-
-      const model = new OpenAIResponsesLanguageModel('gpt-4o', {
-        provider: 'test-provider',
-        url: ({ path }) => `https://api.openai.com/v1${path}`,
-        headers: () => ({}),
-        generateId: mockId(),
-        transformRequestBody: transformFn,
-      });
-
-      await model.doGenerate({
-        prompt: TEST_PROMPT,
-      });
-
-      // Verify transform was called
-      expect(transformFn).toHaveBeenCalledOnce();
-      expect(transformFn).toHaveBeenCalledWith(
-        expect.objectContaining({
-          model: 'gpt-4o',
-          input: [
-            {
-              role: 'user',
-              content: [{ type: 'input_text', text: 'Hello' }],
-            },
-          ],
-        }),
-      );
-
-      // Verify transformed body was sent
-      expect(await server.calls[0].requestBodyJson).toMatchObject({
-        custom_field: 'added-by-transform',
-      });
-    });
-
-    it('should transform request body in doStream when transformRequestBody is provided', async () => {
-      const transformFn = vi.fn((body: Record<string, any>) => ({
-        ...body,
-        custom_field: 'added-by-transform',
-      }));
-
-      prepareTransformStreamResponse();
-
-      const model = new OpenAIResponsesLanguageModel('gpt-4o', {
-        provider: 'test-provider',
-        url: ({ path }) => `https://api.openai.com/v1${path}`,
-        headers: () => ({}),
-        generateId: mockId(),
-        transformRequestBody: transformFn,
-      });
-
-      const { stream } = await model.doStream({
-        prompt: TEST_PROMPT,
-      });
-
-      // Consume the stream
-      await convertReadableStreamToArray(stream);
-
-      // Verify transform was called
-      expect(transformFn).toHaveBeenCalledOnce();
-      expect(transformFn).toHaveBeenCalledWith(
-        expect.objectContaining({
-          model: 'gpt-4o',
-          input: [
-            {
-              role: 'user',
-              content: [{ type: 'input_text', text: 'Hello' }],
-            },
-          ],
-          stream: true,
-        }),
-      );
-
-      // Verify transformed body was sent
-      expect(await server.calls[0].requestBodyJson).toMatchObject({
-        custom_field: 'added-by-transform',
-      });
-    });
-
-    it('should work without transformRequestBody', async () => {
-      prepareTransformJsonResponse();
-
-      const model = new OpenAIResponsesLanguageModel('gpt-4o', {
-        provider: 'test-provider',
-        url: ({ path }) => `https://api.openai.com/v1${path}`,
-        headers: () => ({}),
-        generateId: mockId(),
-        // No transformRequestBody provided
-      });
-
-      await model.doGenerate({
-        prompt: TEST_PROMPT,
-      });
-
-      // Should still work and send the original body
-      const requestBody = await server.calls[0].requestBodyJson;
-      expect(requestBody).toMatchObject({
-        model: 'gpt-4o',
-      });
-      expect(requestBody).not.toHaveProperty('custom_field');
     });
   });
 });
